[*] Smooth estimated time calculation.
git-svn-id: svn://svn.code.sf.net/p/xps2img/code/trunk@1049 f56d0dea-3dc4-4344-980a-dda9636af30f

+++ Xps2ImgUI/Model/ConversionProgressEventArgs.cs
@@ -4,12 +4,12 @@ namespace Xps2ImgUI.Model
 {
     public class ConversionProgressEventArgs: EventArgs
     {
-        public int Percent { get; private set; }
+        public double Percent { get; private set; }
         public int Page { get; private set; }
         public string Pages { get; private set; }
         public string File { get; private set; }
 
-        public ConversionProgressEventArgs(int percent, int page, string pages, string file)
+        public ConversionProgressEventArgs(double percent, int page, string pages, string file)
         {
             Percent = percent;
             Page = page;
